// src/api/back/models/ForumUserDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    UserDTO,
    UserDTOFromJSON,
    UserDTOFromJSONTyped,
    UserDTOToJSON,
} from './';

/**
 * 
 * @export
 * @interface ForumUserDto
 */
export interface ForumUserDto {
    /**
     * 
     * @type {UserDTO}
     * @memberof ForumUserDto
     */
    user: UserDTO;
    /**
     * 
     * @type {string}
     * @memberof ForumUserDto
     */
    mutedUntil: string;
    /**
     * 
     * @type {number}
     * @memberof ForumUserDto
     */
    messages: number;
}

export function ForumUserDtoFromJSON(json: any): ForumUserDto {
    return ForumUserDtoFromJSONTyped(json, false);
}

export function ForumUserDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ForumUserDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'user': UserDTOFromJSON(json['user']),
        'mutedUntil': json['mutedUntil'],
        'messages': json['messages'],
    };
}

export function ForumUserDtoToJSON(value?: ForumUserDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'user': UserDTOToJSON(value.user),
        'mutedUntil': value.mutedUntil,
        'messages': value.messages,
    };
}