// src/api/back/models/RoleSubscriptionEntryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import { Role, RoleFromJSON, RoleFromJSONTyped, RoleToJSON } from "./";

/**
 * 
 * @export
 * @interface RoleSubscriptionEntryDto
 */
export interface RoleSubscriptionEntryDto {
  /**
   *
   * @type {Role}
   * @memberof RoleSubscriptionEntryDto
   */
  role: Role;
  /**
   *
   * @type {number}
   * @memberof RoleSubscriptionEntryDto
   */
  endTime: number;
  /**
   *
   * @type {string}
   * @memberof RoleSubscriptionEntryDto
   */
  steamId: string;
}

export function RoleSubscriptionEntryDtoFromJSON(
  json: any,
): RoleSubscriptionEntryDto {
  return RoleSubscriptionEntryDtoFromJSONTyped(json, false);
}

export function RoleSubscriptionEntryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): RoleSubscriptionEntryDto {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    role: RoleFromJSON(json["role"]),
    endTime: json["end_time"],
    steamId: json["steam_id"],
  };
}

export function RoleSubscriptionEntryDtoToJSON(value?: RoleSubscriptionEntryDto | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    role: RoleToJSON(value.role),
    end_time: value.endTime,
    steam_id: value.steamId,
  };
}