// src/api/back/models/SmurfData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    BanStatusDto,
    BanStatusDtoFromJSON,
    BanStatusDtoFromJSONTyped,
    BanStatusDtoToJSON,
    UserDTO,
    UserDTOFromJSON,
    UserDTOFromJSONTyped,
    UserDTOToJSON,
} from './';

/**
 * 
 * @export
 * @interface SmurfData
 */
export interface SmurfData {
    /**
     * 
     * @type {UserDTO}
     * @memberof SmurfData
     */
    user: UserDTO;
    /**
     * 
     * @type {BanStatusDto}
     * @memberof SmurfData
     */
    ban: BanStatusDto;
}

export function SmurfDataFromJSON(json: any): SmurfData {
    return SmurfDataFromJSONTyped(json, false);
}

export function SmurfDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): SmurfData {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'user': UserDTOFromJSON(json['user']),
        'ban': BanStatusDtoFromJSON(json['ban']),
    };
}

export function SmurfDataToJSON(value?: SmurfData | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'user': UserDTOToJSON(value.user),
        'ban': BanStatusDtoToJSON(value.ban),
    };
}