// src/api/back/models/UserBanSummaryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import {
  BanStatusDto,
  BanStatusDtoFromJSON,
  BanStatusDtoFromJSONTyped,
  BanStatusDtoToJSON,
} from "./";

/**
 *
 * @export
 * @interface UserBanSummaryDto
 */
export interface UserBanSummaryDto {
  /**
   *
   * @type {string}
   * @memberof UserBanSummaryDto
   */
  steamId: string;
  /**
   *
   * @type {BanStatusDto}
   * @memberof UserBanSummaryDto
   */
  banStatus: BanStatusDto;
}

export function UserBanSummaryDtoFromJSON(json: any): UserBanSummaryDto {
  return UserBanSummaryDtoFromJSONTyped(json, false);
}

export function UserBanSummaryDtoFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): UserBanSummaryDto {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    steamId: json["steam_id"],
    banStatus: BanStatusDtoFromJSON(json["banStatus"]),
  };
}

export function UserBanSummaryDtoToJSON(value?: UserBanSummaryDto | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    steam_id: value.steamId,
    banStatus: BanStatusDtoToJSON(value.banStatus),
  };
}